fund_analysis/src/fund_analysis/core/config.py
# Configuration management for Fund Analysis

import yaml
from pathlib import Path

class Config:
    def __init__(self, config_path: str):
        self.config_path = Path(config_path)
        self.config = self._load_config()

    def _load_config(self):
        with open(self.config_path, 'r') as f:
            return yaml.safe_load(f)

    def get(self, key, default=None):
        return self.config.get(key, default)
