fund_analysis/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fund-analysis"
version = "0.1.0"
description = "Fundamental Forecasting Pipeline for Energy Trading"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "xgboost",
    "sqlalchemy",
    "psycopg2",
    "pydantic",
    "click",
    "pyyaml"
]
